remove unused logic and dayjs dep
