+++ frontend/package.json
@@ -15,7 +15,6 @@
   "dependencies": {
     "@vueuse/core": "^14.2.1",
     "butterchurn": "^2.6.7",
-    "dayjs": "^1.11.19",
     "md5": "^2.3.0",
     "vue": "^3.5.29",
     "vue-router": "^5.0.3"
